MOR-92: Moved bar showing/closing out of Application class.
